swtich to using post bulk
